requirements/dev.txt
-r base.txt

ipdb==0.10.3
ipython==6.1.0
django-extensions==1.7.9
django-debug-toolbar==1.8
Werkzeug==0.12.2
flake8==3.3.0